# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_handler.CreateCategoryRequest:
    properties:
      icon_url:
        type: string
      merchant_id:
        example: 1
        type: integer
      name:
        example: 生鲜
        type: string
      parent_id:
        example: 0
        type: integer
      sort_order:
        example: 0
        type: integer
      status:
        example: 1
        type: integer
    required:
    - name
    type: object
...
  internal_handler.UpdateCategoryRequest:
    properties:
      icon_url:
        type: string
      name:
        type: string
      sort_order:
        type: integer
      status:
        type: integer
    type: object
...
  yujixinjiang_backend_internal_query.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  yujixinjiang_backend_internal_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: support@yujixinjiang.com
    name: API Support
  description: 豫记信疆微信小程序后端接口文档，支持在线调试。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 豫记信疆 API
  version: "1.0"
paths:
...
  /admin/categories:
    get:
      parameters:
      - description: 按商家筛选
        in: query
        name: merchant_id
        type: integer
      - description: 0/1
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 分类列表（管理端，含隐藏）
      tags:
      - 管理端-分类
    post:
      consumes:
      - application/json
      parameters:
      - description: 分类信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 创建分类
      tags:
      - 管理端-分类
  /admin/categories/{id}:
    delete:
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 删除分类（逻辑删除）
      tags:
      - 管理端-分类
    get:
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 分类详情
      tags:
      - 管理端-分类
    patch:
      consumes:
      - application/json
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 更新分类
      tags:
      - 管理端-分类
...
  /admin/inventory-usages:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: 按商家筛选
        in: query
        name: merchant_id
        type: integer
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_query.PageResult'
              type: object
      security:
      - BearerAuth: []
      summary: 背包使用记录（管理端）
      tags:
      - 管理端-背包
...
  /admin/users:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: 昵称/手机号
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_query.PageResult'
              type: object
      security:
      - BearerAuth: []
      summary: 用户列表（管理端）
      tags:
      - 管理端-用户
...
securityDefinitions:
  BearerAuth:
    description: '登录后获取的 JWT，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"